+++ ledger-core/network/servicenetwork/status.go
@@ -9,9 +9,9 @@ import (
 	"context"
 	"time"
 
+	"github.com/insolar/assured-ledger/ledger-core/appctl"
 	"github.com/insolar/assured-ledger/ledger-core/insolar/pulsestor"
 	"github.com/insolar/assured-ledger/ledger-core/version"
-	"github.com/insolar/assured-ledger/ledger-core/virtual/pulsemanager"
 )
 
 var startTime time.Time
@@ -24,7 +24,7 @@ func (n *ServiceNetwork) GetNetworkStatus() pulsestor.StatusReply {
 	if err != nil {
 		np = pulsestor.GenesisPulse
 	}
-	reply.Pulse = pulsemanager.ConvertForLegacy(np)
+	reply.Pulse = ConvertForLegacy(np)
 
 	activeNodes := n.NodeKeeper.GetAccessor(np.PulseNumber).GetActiveNodes()
 	workingNodes := n.NodeKeeper.GetAccessor(np.PulseNumber).GetWorkingNodes()
@@ -42,6 +42,29 @@ func (n *ServiceNetwork) GetNetworkStatus() pulsestor.StatusReply {
 
 	return reply
 }
+
+func ConvertForLegacy(pc appctl.PulseChange) (psp pulsestor.Pulse) {
+	pd := pc.Data
+
+	copy(psp.Entropy[:], pd.PulseEntropy[:])
+
+	psp.PulseNumber = pd.PulseNumber
+	ok := false
+
+	if psp.PrevPulseNumber, ok = pd.PulseNumber.TryPrev(pd.PrevPulseDelta); !ok {
+		psp.PrevPulseNumber = pd.PulseNumber
+	}
+
+	if psp.NextPulseNumber, ok = pd.PulseNumber.TryNext(pd.NextPulseDelta); !ok {
+		psp.NextPulseNumber = pd.PulseNumber
+	}
+	psp.EpochPulseNumber = pd.PulseEpoch
+	psp.PulseTimestamp = int64(pd.Timestamp) * int64(time.Second)
+	copy(psp.OriginID[:], pc.PulseOrigin)
+
+	return psp
+}
+
 
 func init() {
 	startTime = time.Now()
